move logic to logic folder and use import
