+++ functions/src/helloWorld.ts
@@ -1,5 +1,5 @@
 import { Request } from 'firebase-functions/lib/providers/https';
-const { firebase, functions } = require('./shared');
+import { firebase, functions } from './shared';
 
 export const helloWorld = functions.https.onRequest((request: Request, response: any) => {
   firebase.logger.info('Hello logs!', { structuredData: true });
