+++ functions/src/index.ts
@@ -1,4 +1,5 @@
-const { helloWorld } = require('./helloWorld');
-const { ingest } = require('./ingest');
+import * as buildQueue from './buildQueue';
+import * as ingest from './ingest';
+import { helloWorld } from './helloWorld';
 
-export { helloWorld, ingest };
+export { helloWorld, ingest, buildQueue };
