+++ functions/src/ingest.ts
@@ -1,31 +1,28 @@
 import { Request } from 'firebase-functions/lib/providers/https';
+import { EventContext } from 'firebase-functions';
+import { getUnityVersionInfo } from './logic/getUnityVersionInfo';
+import { firebase, functions } from './shared';
 
-const { firebase, functions } = require('./shared');
-const { getDocumentFromUrl } = require('./utils/get-document-from-url');
-
-/**
- * Based on https://github.com/BLaZeKiLL/unity-scraper
- */
-export const ingest = functions.https.onRequest(async (request: Request, response: any) => {
+export const intoBrowser = functions.https.onRequest(async (request: Request, response: any) => {
   try {
-    const archiveUrl = 'https://unity3d.com/get-unity/download/archive';
-
-    const document = await getDocumentFromUrl(archiveUrl);
-    const versionInfo = Array.from(
-      document.querySelectorAll('.unityhub') as NodeListOf<HTMLAnchorElement>,
-      (a) => {
-        const link = a.href.replace('unityhub://', '');
-
-        return {
-          version: link.split('/')[0],
-          changeset: link.split('/')[1],
-        };
-      },
-    );
-
+    const versionInfo = await getUnityVersionInfo();
     response.send(versionInfo);
   } catch (err) {
     firebase.logger.error(err);
     response.send('Oops.');
   }
 });
+
+export const intoDb = functions.pubsub
+  .schedule('every 5 minutes')
+  .onRun(async (context: EventContext) => {
+    try {
+      const versionInfo = await getUnityVersionInfo();
+
+      // Todo - ingest in db
+
+      firebase.logger.info(versionInfo);
+    } catch (err) {
+      firebase.logger.error(err);
+    }
+  });
